Improve hard constraint on Boundary Condition (#7)
* Improve hard constraint on Boundary Condition  (#4)

* Add Tony's toy code

* Add missing file

* Create note.tex

* revise encforce bc equation and learning rate

* fix bug: setting frequency

* AD: retain graph

* remove g0 g1

* enforce BC by linear

* fix: use linear scale BC enforcement

* remove retain_graph=True followed by https://docs.pytorch.org/docs/stable/generated/torch.autograd.grad.html

---------


* minor change

* code cleanup

---------

+++ pinn/utils.py
@@ -44,7 +44,7 @@ def parse_args(args=None):
                         help="Lower bound of the 1D domain.")
     parser.add_argument('--bx', type=float, default=1.0,
                         help="Upper bound of the 1D domain.")
-    parser.add_argument('--high_freq', type=float, default=16,
+    parser.add_argument('--high_freq', type=int, default=16,
                         help="Highest frequency used in the PDE solution (PDE 1).")
     parser.add_argument('--gamma', type=float, default=0,
                         help="Coefficient \gamma in the PDE: -u_xx + \gamma u = f.")
